+++ modulos/design_system_module/lib/src/mixins/ui/loader/loader_mixin.dart
@@ -9,15 +9,19 @@ mixin LoaderMixin on GetxController {
       statusLoad,
       (loading) {
         if (loading) {
-          WidgetsBinding.instance.addPostFrameCallback((duration) async {
-            await Get.dialog(
-              const Center(
-                child: CircularProgressIndicator(),
-              ),
-            );
-          });
+          Get.dialog(
+            const Center(
+              child: CircularProgressIndicator(),
+            ),
+          );
         } else {
           Get.back();
+        }
+        if (!loading) {
+          bool? dialogOpen = Get.isDialogOpen;
+          if (dialogOpen != null && dialogOpen) {
+            Navigator.pop(Get.context!);
+          }
         }
       },
     );