+++ modulos/remessas_module/lib/src/ui/widgets/bodyops/bodyupload_csv_widget.dart
@@ -102,7 +102,7 @@ _todasRemessasList() {
                         designSystemController.iconButtonPrint(
                           filtro: remessaModel,
                         ),
-                        SizedBox(
+                        const SizedBox(
                           width: 15,
                         ),
                         designSystemController.iconDownloadXlsx(