+++ modulos/splash_module/lib/src/ui/splash_page.dart
@@ -7,53 +7,10 @@ class SplashPage extends GetView<SplashController> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.primary,
-      body: Center(
-        child: Column(
-          children: [
-            const Text(
-              "Splash - Bem Vindo ao Sistema..",
-              style: TextStyle(
-                fontSize: 30.0,
-                color: Colors.white,
-              ),
-            ),
-            const Text(
-              "teste",
-              style: TextStyle(
-                fontSize: 30.0,
-                color: Colors.white,
-              ),
-            ),
-            Text(
-              Routes.initial.caminho,
-              style: const TextStyle(
-                fontSize: 30.0,
-                color: Colors.white,
-              ),
-            ),
-            Center(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  const Text(
-                    'You have pushed the button this many times:',
-                  ),
-                  Obx(() => Text(
-                        '${controller.counter}',
-                        style: Theme.of(context).textTheme.headline4,
-                      )),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: controller.incrementCounter,
-        tooltip: 'Increment',
-        child: const Icon(Icons.add),
+    return Container(
+      color: Colors.grey,
+      child: const Center(
+        child: CircularProgressIndicator(),
       ),
     );
   }