+++ modulos/design_system_module/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* TODO: 
+1 - Correção de bug no loading.
+2 - Cache de imagem da base do protocolo
+
 ## 0.0.1
 
 * TODO: Describe initial release.